Get basic information about Github repo

Get all repo info offered by the main type of Gihub API requests.

github_repo_info(repo.url, user = NA, repo = NA)

Arguments

user
character username

repo
character repo name

Value

List of Github repo attributes

References

For detailed info see: https://developer.github.com/v3/

Examples

github_repo_info("github.com/qinwf/awesome-r")

#> $id
#> [1] 22003158
#> 
#> $node_id
#> [1] "MDEwOlJlcG9zaXRvcnkyMjAwMzE1OA=="
#> 
#> $name
#> [1] "awesome-R"
#> 
#> $full_name
#> [1] "qinwf/awesome-R"
#> 
#> $private
#> [1] FALSE
#> 
#> $owner
#> $owner$login
#> [1] "qinwf"
#> 
#> $owner$id
#> [1] 7472151
#> 
#> $owner$node_id
#> [1] "MDQ6VXNlcjc0NzIxNTE="
#> 
#> $owner$avatar_url
#> [1] "https://avatars1.githubusercontent.com/u/7472151?v=4"
#> 
#> $owner$gravatar_id
#> [1] ""
#> 
#> $owner$url
#> [1] "https://api.github.com/users/qinwf"
#> 
#> $owner$html_url
#> [1] "https://github.com/qinwf"
#> 
#> $owner$followers_url
#> [1] "https://api.github.com/users/qinwf/followers"
#> 
#> $owner$following_url
#> [1] "https://api.github.com/users/qinwf/following{/other_user}"
#> 
#> $owner$gists_url
#> [1] "https://api.github.com/users/qinwf/gists{/gist_id}"
#> 
#> $owner$starred_url
#> [1] "https://api.github.com/users/qinwf/starred{/owner}{/repo}"
#> 
#> $owner$subscriptions_url
#> [1] "https://api.github.com/users/qinwf/subscriptions"
#> 
#> $owner$organizations_url
#> [1] "https://api.github.com/users/qinwf/orgs"
#> 
#> $owner$repos_url
#> [1] "https://api.github.com/users/qinwf/repos"
#> 
#> $owner$events_url
#> [1] "https://api.github.com/users/qinwf/events{/privacy}"
#> 
#> $owner$received_events_url
#> [1] "https://api.github.com/users/qinwf/received_events"
#> 
#> $owner$type
#> [1] "User"
#> 
#> $owner$site_admin
#> [1] FALSE
#> 
#> 
#> $html_url
#> [1] "https://github.com/qinwf/awesome-R"
#> 
#> $description
#> [1] "A curated list of awesome R packages, frameworks and software."
#> 
#> $fork
#> [1] FALSE
#> 
#> $url
#> [1] "https://api.github.com/repos/qinwf/awesome-R"
#> 
#> $forks_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/forks"
#> 
#> $keys_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/keys{/key_id}"
#> 
#> $collaborators_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/collaborators{/collaborator}"
#> 
#> $teams_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/teams"
#> 
#> $hooks_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/hooks"
#> 
#> $issue_events_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/issues/events{/number}"
#> 
#> $events_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/events"
#> 
#> $assignees_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/assignees{/user}"
#> 
#> $branches_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/branches{/branch}"
#> 
#> $tags_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/tags"
#> 
#> $blobs_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/git/blobs{/sha}"
#> 
#> $git_tags_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/git/tags{/sha}"
#> 
#> $git_refs_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/git/refs{/sha}"
#> 
#> $trees_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/git/trees{/sha}"
#> 
#> $statuses_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/statuses/{sha}"
#> 
#> $languages_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/languages"
#> 
#> $stargazers_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/stargazers"
#> 
#> $contributors_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/contributors"
#> 
#> $subscribers_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/subscribers"
#> 
#> $subscription_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/subscription"
#> 
#> $commits_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/commits{/sha}"
#> 
#> $git_commits_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/git/commits{/sha}"
#> 
#> $comments_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/comments{/number}"
#> 
#> $issue_comment_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/issues/comments{/number}"
#> 
#> $contents_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/contents/{+path}"
#> 
#> $compare_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/compare/{base}...{head}"
#> 
#> $merges_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/merges"
#> 
#> $archive_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/{archive_format}{/ref}"
#> 
#> $downloads_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/downloads"
#> 
#> $issues_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/issues{/number}"
#> 
#> $pulls_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/pulls{/number}"
#> 
#> $milestones_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/milestones{/number}"
#> 
#> $notifications_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/notifications{?since,all,participating}"
#> 
#> $labels_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/labels{/name}"
#> 
#> $releases_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/releases{/id}"
#> 
#> $deployments_url
#> [1] "https://api.github.com/repos/qinwf/awesome-R/deployments"
#> 
#> $created_at
#> [1] "2014-07-19T06:02:47Z"
#> 
#> $updated_at
#> [1] "2019-09-14T20:27:04Z"
#> 
#> $pushed_at
#> [1] "2019-09-12T05:40:19Z"
#> 
#> $git_url
#> [1] "git://github.com/qinwf/awesome-R.git"
#> 
#> $ssh_url
#> [1] "git@github.com:qinwf/awesome-R.git"
#> 
#> $clone_url
#> [1] "https://github.com/qinwf/awesome-R.git"
#> 
#> $svn_url
#> [1] "https://github.com/qinwf/awesome-R"
#> 
#> $homepage
#> [1] "https://awesome-r.com"
#> 
#> $size
#> [1] 1123
#> 
#> $stargazers_count
#> [1] 3670
#> 
#> $watchers_count
#> [1] 3670
#> 
#> $language
#> [1] "R"
#> 
#> $has_issues
#> [1] TRUE
#> 
#> $has_projects
#> [1] TRUE
#> 
#> $has_downloads
#> [1] TRUE
#> 
#> $has_wiki
#> [1] TRUE
#> 
#> $has_pages
#> [1] TRUE
#> 
#> $forks_count
#> [1] 1225
#> 
#> $mirror_url
#> NULL
#> 
#> $archived
#> [1] FALSE
#> 
#> $disabled
#> [1] FALSE
#> 
#> $open_issues_count
#> [1] 13
#> 
#> $license
#> NULL
#> 
#> $forks
#> [1] 1225
#> 
#> $open_issues
#> [1] 13
#> 
#> $watchers
#> [1] 3670
#> 
#> $default_branch
#> [1] "master"
#> 
#> $network_count
#> [1] 1225
#> 
#> $subscribers_count
#> [1] 387
#>